IizakaEmpire/requirements.txt
aiohttp==3.6.2
asgiref==3.2.10
async-timeout==3.0.1
attrs==20.2.0
autopep8==1.5.4
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
Django==3.1.2
django-cleanup==5.1.0
django-markdownx==3.0.1
django-silk==4.1.0
django-storages==1.10.1
google-api-core==1.22.3
google-auth==1.22.0
google-cloud-core==1.4.2
google-cloud-storage==1.31.2
google-crc32c==1.0.0
google-resumable-media==1.0.0
googleapis-common-protos==1.52.0
gprof2dot==2019.11.30
idna==2.10
Jinja2==2.11.2
Markdown==3.2.2
MarkupSafe==1.1.1
multidict==4.7.6
mysqlclient==2.0.1
Pillow==7.2.0
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
Pygments==2.7.2
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
rsa==4.6
six==1.15.0
sqlparse==0.3.1
toml==0.10.2
urllib3==1.25.10
uWSGI==2.0.19.1
yarl==1.6.0
djangorestframework==3.12.2
django-filter==2.4.0